functions/link2_consume/requirements.txt
azure-core==1.9.0
azure-storage-file-share==12.3.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
dataclasses==0.6
defusedxml==0.6.0
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-secret-manager==2.1.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.34.0
idna==2.10
isodate==0.6.0
libcst==0.3.16
msrest==0.6.19
mypy-extensions==0.4.3
oauthlib==3.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
